webapi/tct-notification-tizen-tests/notification/NotificationManager_get.html

<!DOCTYPE html>
<!--
Copyright (c) 2012 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Lei, ZhanX <zhanx.lei@intel.com>
        Zhang, GeX <gex.zhang@intel.com>
        Andrzej Krolikowski <a.krolikowsk@samsung.com>

-->
<html>
<head>
<title>NotificationManager_get</title>
<meta charset="utf-8"/>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="support/notification_common.js"></script>
</head>

<body>
<div id="log"></div>
<script type="text/javascript">
//==== TEST: NotificationManager_get
//==== LABEL Check if the parameter NotificationManager which in get method ok
//==== PRIORITY P2
//==== SPEC Tizen Web API:User Interface:Notification:NotificationManager:get M
//==== SPEC_URL https://developer.tizen.org/help/topic/org.tizen.web.device.apireference/tizen/notification.html
//==== TEST_CRITERIA MR MMINA

var notification, statusType = "PROGRESS", progressTypeToSet = "PERCENTAGE",
    notificationToPost, postedTime, i;
test(function () {
    notificationDict.progressType = progressTypeToSet;
    notificationToPost = new tizen.StatusNotification(statusType, titleToSet, notificationDict);
    try {
        tizen.notification.post(notificationToPost);
        postedTime = notificationToPost.postedTime;

        notification = tizen.notification.get(notificationToPost.id);
        assert_true(notification instanceof tizen.StatusNotification, "Incorrect returned value from get method.");

        assert_equals(notification.type, "STATUS", "type is incorrect.");
        assert_equals(notification.postedTime.toJSON(), postedTime.toJSON(), "postedTime is incorrect.");

        assert_equals(notification.title, titleToSet, "title is incorrect.");
        assert_equals(notification.content, contentToSet, "content is incorrect.");
        assert_equals(notification.iconPath, iconPathToSet, "iconPath is incorrect.");
        assert_equals(notification.soundPath, soundPathToSet, "soundPath is incorrect.");
        assert_equals(notification.vibration, vibrationToSet, "vibration is incorrect.");

        assert_equals(notification.appId, appIdToSet, "appId is incorrect.");
        assert_equals(notification.appControl.operation, applicationControl.operation,
                "appControl.operation is incorrect.");
        assert_equals(notification.appControl.uri, applicationControl.uri, "appControl.uri is incorrect.");
        assert_equals(notification.appControl.mime, applicationControl.mime, "appControl.mime is incorrect.");
        assert_equals(notification.appControl.category, applicationControl.category,
                "appControl.category is incorrect.");
        assert_array_equals(notification.appControl.data, applicationControl.data,
                "appControl.data is incorrect.");

        assert_equals(notification.id, notificationToPost.id, "id is incorrect.");
        assert_equals(notification.progressValue, progressValueToSet, "progressValue is incorrect.");
        assert_equals(notification.progressType, progressTypeToSet, "progressType is incorrect.");
        assert_equals(notification.statusType, statusType, "statusType is incorrect.");

        assert_equals(notification.subIconPath, subIconPathToSet, "subIconPath is incorrect.");
        assert_equals(notification.number, numberToSet, "number is incorrect.");
        assert_equals(notification.backgroundImagePath, backgroundImagePathToSet, "backgroundImagePath is incorrect.");

        assert_array_equals(notification.thumbnails, thumbnailsToSet, "thumbnails attribute.");
        assert_equals(notification.detailInfo.length, detailInfoToSet.length, "detailInfo attribute length.");
        for (i = 0; i < detailInfoToSet.length; i++) {
            assert_equals(notification.detailInfo[i].mainText, detailInfoToSet[i].mainText,
                "detailInfo.mainText attribute for index = " + i + ".");
            assert_equals(notification.detailInfo[i].subText, detailInfoToSet[i].subText,
                "detailInfo.subText attribute for index = " + i + ".");
        }
        assert_own_property(notification, "ledColor", "notification");
        assert_equals(notification.ledOnPeriod, 0, "notification.ledOnPeriod");
        assert_equals(notification.ledOffPeriod, 0, "notification.ledOffPeriod");
    } finally {
        tizen.notification.removeAll();
    }
}, document.title);

</script>
</body>
</html>
